requirements.txt
filelock==3.17.0
fsspec==2025.2.0
iniconfig==2.0.0
Jinja2==3.1.5
line_profiler==4.2.0
llvmlite==0.44.0
MarkupSafe==3.0.2
mpi4py==4.0.3
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
PyYAML==6.0.2
scipy==1.15.2
sympy==1.13.1
torch==2.6.0
triton==3.2.0
typing_extensions==4.12.2
